Simplify the condition to skip non-running containers
The logic in this conditional was broken for Fargate containers, as the check on container.Status != "running" was removed.

I took some time to simplify the if, as it was very difficult to read. The logic to identify containers that are not running is:

their state isn't running (case insensitive, covers classic cgroups containers)
their status isn't running (case insensitive, covers ECS containers)
their status doesn't start with up (case insensitive, covers SystemD containers)